+++ src/settings/internal/bool_setting_value.h
