+++ src/settings/internal/double_setting_value.h
