+++ src/settings/internal/setting_value.h
